Fixed RandomLetterGenerator so each letter is unique in the output array.

+++ utilities/RandomLetterGenerator.h
@@ -1,14 +1,16 @@
 #ifndef RANDOMLETTERGENERATOR_H
 #define RANDOMLETTERGENERATOR_H
 
+#include <vector>
 #include <cstdlib>
+#include <stdexcept>
 using namespace std;
 
 class RandomLetterGenerator
 {
 
 public:
-    static char* makeRandomUniqueLetterArray(int size);
+    static char* makeRandomUniqueLetterArray(size_t size);
 
 };
